chore: move from rollup to rslib
