+++ packages/next-admin/package.json
@@ -29,8 +29,8 @@
   "homepage": "https://next-admin.js.org/",
   "license": "MIT",
   "scripts": {
-    "build": "rollup -c --configPlugin typescript && cp src/theme.css dist/theme.css",
-    "dev": "rollup -c --configPlugin typescript --watch",
+    "build": "rslib build",
+    "dev": "rslib build --watch",
     "lint": "eslint \"**/*.{ts,tsx}\"",
     "test": "vitest",
     "typecheck": "tsc --noEmit",
@@ -206,7 +206,8 @@
   "devDependencies": {
     "@premieroctet/next-admin-json-schema": "workspace:*",
     "@remix-run/react": "catalog:remix",
-    "@rollup/plugin-typescript": "^12.1.2",
+    "@rsbuild/plugin-react": "^1.3.2",
+    "@rslib/core": "^0.9.2",
     "@tanstack/react-router": "catalog:tanstack-start",
     "@testing-library/react": "^14.1.2",
     "@types/body-parser": "^1.19.2",
@@ -232,8 +233,6 @@
     "lodash.debounce": "^4.0.8",
     "next": "^15.3.1",
     "postcss": "^8.4.21",
-    "rollup": "^4.40.1",
-    "rollup-preserve-directives": "^1.1.3",
     "tailwindcss": "^3.2.7",
     "ts-node": "^10.9.1",
     "tsconfig": "workspace:*",
